Extract common ingest pipeline code
